refactor: Use Go types in tq.go

+++ src/aprs_tt.go
@@ -504,7 +504,7 @@ func aprs_tt_sequence(channel int, msg string) {
 		return
 	}
 
-	tq_append(C.int(channel), TQ_PRIO_0_HI, pp)
+	tq_append(channel, TQ_PRIO_0_HI, pp)
 } /* end aprs_tt_sequence */
 
 /*------------------------------------------------------------------
