+++ src/ax25_link.go
@@ -989,7 +989,7 @@ func dl_disconnect_request(E *dlq_item_t) {
 		var p1 = 1
 		var nopid0 = 0
 		var pp15 = ax25_u_frame(S.addrs, S.num_addr, cr_cmd, frame_type_U_DISC, p1, nopid0, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp15)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp15)
 
 		STOP_T1(S) // started in establish_data_link.
 		STOP_T3(S) // probably don't need.
@@ -1013,7 +1013,7 @@ func dl_disconnect_request(E *dlq_item_t) {
 			var nopid = 0 // PID applies only to I and UI frames.
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, cr, frame_type_U_DM, p, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_0_HI, pp) // HI means expedited.
+			lm_data_request(S.channel, TQ_PRIO_0_HI, pp) // HI means expedited.
 
 			// Erratum: Shouldn't we inform the user when going to disconnected state?
 			// Notifying the application, here, is my own enhancement.
@@ -1036,7 +1036,7 @@ func dl_disconnect_request(E *dlq_item_t) {
 		var nopid = 0
 
 		var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, frame_type_U_DISC, p, nopid, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 		STOP_T3(S)
 		START_T1(S)
@@ -1363,7 +1363,7 @@ func data_request_good_size(S *ax25_dlsm_t, txdata *cdata_t) {
 		if (!S.peer_receiver_busy) &&
 			WITHIN_WINDOW_SIZE(S) {
 			S.acknowledge_pending = true
-			lm_seize_request(C.int(S.channel))
+			lm_seize_request(S.channel)
 		}
 
 	default:
@@ -2160,7 +2160,7 @@ func lm_data_indication(E *dlq_item_t) {
 		WITHIN_WINDOW_SIZE(S) {
 
 		//S.acknowledge_pending = 1;
-		lm_seize_request(C.int(S.channel))
+		lm_seize_request(S.channel)
 	}
 
 } /* end lm_data_indication */
@@ -2241,7 +2241,7 @@ func i_frame(S *ax25_dlsm_t, cr cmdres_t, p int, nr int, ns int, pid int, info [
 			var nopid = 0 // PID applies only for I and UI frames.
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 	case state_1_awaiting_connection, state_5_awaiting_v22_connection:
@@ -2258,7 +2258,7 @@ func i_frame(S *ax25_dlsm_t, cr cmdres_t, p int, nr int, ns int, pid int, info [
 			var nopid = 0 // PID applies only for I and UI frames.
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 	case state_3_connected, state_4_timer_recovery:
@@ -2346,7 +2346,7 @@ func i_frame(S *ax25_dlsm_t, cr cmdres_t, p int, nr int, ns int, pid int, info [
 							#else // 2006 version - states 3 & 4 the same.
 						*/
 
-						lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+						lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 						// #endif
 						S.acknowledge_pending = false
@@ -2546,7 +2546,7 @@ func i_frame_continued(S *ax25_dlsm_t, p int, ns int, pid int, info []byte) {
 			var cr = cr_res // response with F set to 1.
 
 			var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RR, S.modulo, nr, f, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			S.acknowledge_pending = false
 		} else if !S.acknowledge_pending {
 
@@ -2559,7 +2559,7 @@ func i_frame_continued(S *ax25_dlsm_t, p int, ns int, pid int, info []byte) {
 			// When that event arrives, we check acknowledge_pending and send something
 			// to be determined later.
 
-			lm_seize_request(C.int(S.channel))
+			lm_seize_request(S.channel)
 		}
 	} else if S.reject_exception {
 
@@ -2574,7 +2574,7 @@ func i_frame_continued(S *ax25_dlsm_t, p int, ns int, pid int, info []byte) {
 			var cr = cr_res // response with F set to 1.
 
 			var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RR, S.modulo, nr, f, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			S.acknowledge_pending = false
 		}
 	} else if S.srej_enable == srej_none {
@@ -2599,7 +2599,7 @@ func i_frame_continued(S *ax25_dlsm_t, p int, ns int, pid int, info []byte) {
 		}
 
 		var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_REJ, S.modulo, nr, f, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 		S.acknowledge_pending = false
 	} else {
@@ -2645,7 +2645,7 @@ func i_frame_continued(S *ax25_dlsm_t, p int, ns int, pid int, info []byte) {
 				var nr = S.vr
 
 				var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RNR, S.modulo, nr, f, nil)
-				lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+				lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			} else if S.rxdata_by_ns[AX25MODULO(ns-1, S.modulo)] == nil {
 
 				// Ask for missing frames when we don't have N(S)-1 in the receive buffer.
@@ -3016,7 +3016,7 @@ func send_srej_frames(S *ax25_dlsm_t, resend []int, count int, allow_f1 bool) {
 		}
 
 		var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_SREJ, S.modulo, nr, _f, info)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		return
 	}
 
@@ -3048,7 +3048,7 @@ func send_srej_frames(S *ax25_dlsm_t, resend []int, count int, allow_f1 bool) {
 		}
 
 		var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_SREJ, S.modulo, nr, _f, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 	}
 
 } /* end send_srej_frames */
@@ -3109,7 +3109,7 @@ func rr_rnr_frame(S *ax25_dlsm_t, ready bool, cr cmdres_t, pf int, nr int) {
 			var f = pf
 			var nopid int = 0 // PID only for I and UI frames.
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 	case state_1_awaiting_connection, state_5_awaiting_v22_connection:
@@ -3126,7 +3126,7 @@ func rr_rnr_frame(S *ax25_dlsm_t, ready bool, cr cmdres_t, pf int, nr int) {
 			var nopid int = 0 // PID applies only for I and UI frames.
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 		// Erratum: We have a disagreement here between original and 2006 version.
@@ -3354,7 +3354,7 @@ func rej_frame(S *ax25_dlsm_t, cr cmdres_t, pf int, nr int) {
 			var nopid int = 0 // PID is only for I and UI.
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 	case state_1_awaiting_connection, state_5_awaiting_v22_connection:
@@ -3368,7 +3368,7 @@ func rej_frame(S *ax25_dlsm_t, cr cmdres_t, pf int, nr int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 
 		// Erratum: We have a disagreement here between original and 2006 version.
@@ -3781,7 +3781,7 @@ func resend_for_srej(S *ax25_dlsm_t, nr int, info []byte) int {
 	if txdata != nil {
 		var pp = ax25_i_frame(S.addrs, S.num_addr, cr, S.modulo, i_frame_nr, i_frame_ns, p, txdata.pid, txdata.data[:txdata.len])
 		// dw_printf ("calling lm_data_request for I frame, %s line %d\n", __func__, __LINE__);
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		num_resent++
 	} else {
 		text_color_set(DW_COLOR_ERROR)
@@ -3812,7 +3812,7 @@ func resend_for_srej(S *ax25_dlsm_t, nr int, info []byte) int {
 		txdata = S.txdata_by_ns[i_frame_ns]
 		if txdata != nil {
 			var pp = ax25_i_frame(S.addrs, S.num_addr, cr, S.modulo, i_frame_nr, i_frame_ns, p, txdata.pid, txdata.data[:txdata.len])
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			num_resent++
 		} else {
 			text_color_set(DW_COLOR_ERROR)
@@ -3893,7 +3893,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 		var nopid int = 0 // PID is only for I and UI.
 
 		var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 		clear_exception_conditions(S)
 
@@ -3928,7 +3928,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			enter_new_state(S, state_5_awaiting_v22_connection)
 		} else { // SABM - respond with UA.
 
@@ -3941,7 +3941,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			// stay in state 1.
 		}
 
@@ -3953,7 +3953,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			// stay in state 5
 		} else { // SABM, respond with UA, enter state 1
 			var res = cr_res
@@ -3961,7 +3961,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			enter_new_state(S, state_1_awaiting_connection)
 		}
 
@@ -3976,7 +3976,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_0_HI, pp) // expedited
+			lm_data_request(S.channel, TQ_PRIO_0_HI, pp) // expedited
 			// stay in state 2.
 		}
 
@@ -3988,7 +3988,7 @@ func sabm_e_frame(S *ax25_dlsm_t, extended bool, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			// State 3 & 4 handling are the same except for this one difference.
 			if S.state == state_4_timer_recovery {
@@ -4068,7 +4068,7 @@ func disc_frame(S *ax25_dlsm_t, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 		}
 		// keep current state, 0, 1, or 5.
 
@@ -4080,7 +4080,7 @@ func disc_frame(S *ax25_dlsm_t, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_0_HI, pp) // expedited
+			lm_data_request(S.channel, TQ_PRIO_0_HI, pp) // expedited
 		}
 		// keep current state, 2.
 
@@ -4094,7 +4094,7 @@ func disc_frame(S *ax25_dlsm_t, p int) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_UA, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			// dl disconnect *indication*
 			text_color_set(DW_COLOR_INFO)
@@ -4608,7 +4608,7 @@ func ui_frame(S *ax25_dlsm_t, cr cmdres_t, pf int) {
 				var nopid int = 0 // PID applies only for I and UI frames.
 
 				var pp = ax25_u_frame(S.addrs, S.num_addr, r, frame_type_U_DM, pf, nopid, nil)
-				lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+				lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			}
 
 		case state_3_connected, state_4_timer_recovery:
@@ -4696,7 +4696,7 @@ func xid_frame(S *ax25_dlsm_t, cr cmdres_t, pf int, info []byte) {
 					var nopid int = 0
 					var f int = -1
 					var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_XID, f, nopid, xinfo)
-					lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+					lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 				}
 			} else {
 				text_color_set(DW_COLOR_ERROR)
@@ -4737,7 +4737,7 @@ func xid_frame(S *ax25_dlsm_t, cr cmdres_t, pf int, info []byte) {
 				   	          packet_t pp;
 
 				   	          pp = ax25_u_frame (S.addrs, S.num_addr, cmd, frame_type_U_TEST, p, nopid, (unsigned char *)info, (int)strlen(info));
-				   	          lm_data_request (C.int(S.channel), TQ_PRIO_1_LO, pp);
+				   	          lm_data_request (S.channel, TQ_PRIO_1_LO, pp);
 				   	        }
 				   #endif
 				*/
@@ -4794,7 +4794,7 @@ func test_frame(S *ax25_dlsm_t, cr cmdres_t, pf int, info []byte) {
 
 	if cr == cr_cmd {
 		var pp = ax25_u_frame(S.addrs, S.num_addr, res, frame_type_U_TEST, f, nopid, info)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 	}
 } /* end test_frame */
 
@@ -4922,7 +4922,7 @@ func t1_expiry(S *ax25_dlsm_t) {
 				_s = frame_type_U_SABME
 			}
 			var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, _s, p, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			select_t1_value(S)
 			START_T1(S)
 			// Keep same state.
@@ -4946,7 +4946,7 @@ func t1_expiry(S *ax25_dlsm_t) {
 			}
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, frame_type_U_DISC, p, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			select_t1_value(S)
 			START_T1(S)
 			// stay in same state
@@ -5000,7 +5000,7 @@ func t1_expiry(S *ax25_dlsm_t) {
 			var nopid int = 0
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, cr, frame_type_U_DM, f, nopid, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			enter_new_state(S, state_0_disconnected)
 		} else {
@@ -5116,7 +5116,7 @@ func tm201_expiry(S *ax25_dlsm_t) {
 			var xinfo = xid_encode(&param, cmd)
 
 			var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, frame_type_U_XID, p, nopid, xinfo)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			START_TM201(S)
 		}
@@ -5187,7 +5187,7 @@ func establish_data_link(S *ax25_dlsm_t) {
 	}
 
 	var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, frameType, p, nopid, nil)
-	lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+	lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 	STOP_T3(S)
 	START_T1(S)
@@ -5285,7 +5285,7 @@ func transmit_enquiry(S *ax25_dlsm_t) {
 	}
 	var pp = ax25_s_frame(S.addrs, S.num_addr, cmd, ft, S.modulo, nr, p, nil)
 
-	lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+	lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 	S.acknowledge_pending = false
 	START_T1(S)
@@ -5353,7 +5353,7 @@ func enquiry_response(S *ax25_dlsm_t, frame_type ax25_frame_type_t, f int) {
 			// I'm busy.
 
 			var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RNR, S.modulo, nr, f, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			S.acknowledge_pending = false // because we sent N(R) from V(R).
 		} else if S.srej_enable == srej_single || S.srej_enable == srej_multi {
@@ -5402,7 +5402,7 @@ func enquiry_response(S *ax25_dlsm_t, frame_type ax25_frame_type_t, f int) {
 				// Not waiting for fill in of missing frames.		X.25 2.4.6.11 c)
 
 				var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RR, S.modulo, nr, f, nil)
-				lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+				lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 				S.acknowledge_pending = false
 			}
@@ -5421,7 +5421,7 @@ func enquiry_response(S *ax25_dlsm_t, frame_type ax25_frame_type_t, f int) {
 			}
 
 			var pp = ax25_s_frame(S.addrs, S.num_addr, cr, frame_type_S_RR, S.modulo, nr, f, nil)
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			S.acknowledge_pending = false
 		}
@@ -5435,7 +5435,7 @@ func enquiry_response(S *ax25_dlsm_t, frame_type ax25_frame_type_t, f int) {
 			_r = frame_type_S_RNR
 		}
 		var pp = ax25_s_frame(S.addrs, S.num_addr, cr, _r, S.modulo, nr, f, nil)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 		S.acknowledge_pending = false
 	}
@@ -5521,7 +5521,7 @@ func invoke_retransmission(S *ax25_dlsm_t, nr_input int) {
 			var pp = ax25_i_frame(S.addrs, S.num_addr, cr, S.modulo, nr, ns, p,
 				S.txdata_by_ns[ns].pid, S.txdata_by_ns[ns].data[:S.txdata_by_ns[ns].len])
 
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 			// Keep it around in case we need to send again.
 
 			sent_count++
@@ -5965,7 +5965,7 @@ func i_frame_pop_off_queue(S *ax25_dlsm_t) {
 				// dw_printf ("calling lm_data_request for I frame, %s line %d\n", __func__, __LINE__);
 			}
 
-			lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+			lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 			// Stash in sent array in case it gets lost and needs to be sent again.
 
@@ -6096,7 +6096,7 @@ func mdl_negotiate_request(S *ax25_dlsm_t) {
 		var p int = 1
 		var nopid int = 0
 		var pp = ax25_u_frame(S.addrs, S.num_addr, cmd, frame_type_U_XID, p, nopid, xinfo)
-		lm_data_request(C.int(S.channel), TQ_PRIO_1_LO, pp)
+		lm_data_request(S.channel, TQ_PRIO_1_LO, pp)
 
 		S.mdl_rc = 0
 		START_TM201(S)
